add setDF method for macos
