+++ utils_linux.go
@@ -1,0 +1,40 @@
+//go:build linux
+
+package qotp
+
+import (
+	"log/slog"
+	"net"
+
+	"golang.org/x/sys/unix"
+)
+
+// based on: https://github.com/quic-go/quic-go/blob/d540f545b0b70217220eb0fbd5278ece436a7a20/sys_conn_df_linux.go#L15
+func setDontFragment(conn *net.UDPConn) error {
+	rawConn, err := conn.SyscallConn()
+	if err != nil {
+		return err
+	}
+
+	var errDFIPv4, errDFIPv6 error
+	if err := rawConn.Control(func(fd uintptr) {
+		errDFIPv4 = unix.SetsockoptInt(int(fd), unix.IPPROTO_IP, unix.IP_MTU_DISCOVER, unix.IP_PMTUDISC_DO)
+		errDFIPv6 = unix.SetsockoptInt(int(fd), unix.IPPROTO_IPV6, unix.IPV6_MTU_DISCOVER, unix.IPV6_PMTUDISC_DO)
+	}); err != nil {
+		return err
+	}
+
+	switch {
+	case errDFIPv4 == nil && errDFIPv6 == nil:
+		slog.Info("setting DF for IPv4 and IPv6")
+		// TODO: expose this and don't probe for higher MTU when not DF not supported
+	case errDFIPv4 == nil && errDFIPv6 != nil:
+		slog.Info("setting DF for IPv4 only")
+	case errDFIPv4 != nil && errDFIPv6 == nil:
+		slog.Info("setting DF for IPv6 only")
+	case errDFIPv4 != nil && errDFIPv6 != nil:
+		slog.Error("setting DF failed for both IPv4 and IPv6")
+	}
+
+	return nil
+}
